feat: complete exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -46,6 +46,16 @@ blogRoute.post('/', async (request, response) => {
 })
 
 blogRoute.delete('/:id', async (request, response) => {
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id)
+		return response.status(401).json({ error: 'token invalid' })
+
+	const blog = await Blog.findById(request.params.id)
+
+	if (blog.user.toString() !== decodedToken.id.toString())
+		return response.status(401).json({ error: 'token invalid' })
+
 	await Blog.findByIdAndRemove(request.params.id)
 	response.status(204).end()
 })
